+++ module-services/service-cellular/service-cellular/CellularRequestHandler.hpp
@@ -23,4 +23,5 @@ class CellularRequestHandler : public RequestHandler
 
   private:
     ServiceCellular &cellular;
+    void sendActionMessage(bool result);
 };