+++ changelog.md
@@ -4,6 +4,7 @@
 
 * `[cellular]` Integration with basic flow for SIM card (cellular<>GUI)
 * `[file indexer db]` Added UT fot File Indexer DB
+* `[cellular]` Added MMI convertible to action messages.
 
 ### Fixed